Moved login component to pages folder
